after first quick test:
- made race tab visible at startup (RaceVisible = true)
- disabled Icons button

Note that In FR01  I have removed components,
but most of the button-click code is still there in app.component.ts,
much the same as in full app  fr07 - not yet posted.

This is the point in time where I have Just completed the conversion from single app folder to a workspace with libraries and multiple apps.

ng serve fr01 -o is supposed to work.
Tested on Windows 10 with Chrome.

Please ignore the tests at this stage.
and I have not configured lint yet.

We have an initial version!